+++ src/components/App/App.jsx
@@ -12,6 +12,7 @@ import Nav from '../Nav/Nav';
 import Footer from '../Footer/Footer';
 
 import ProtectedRoute from '../ProtectedRoute/ProtectedRoute';
+import MaterialTheme from '../MaterialTheme/MaterialTheme';
 
 import AboutPage from '../AboutPage/AboutPage';
 import UserPage from '../UserPage/UserPage';
@@ -28,7 +29,7 @@ import HowItWorks from '../HowItWorks/HowItWorks';
 import EditProfile from '../EditProfile/EditProfile';
 
 import './App.css';
-import { createTheme, ThemeProvider } from '@mui/material/styles';
+import { ThemeProvider } from '@mui/material/styles';
 
 
 function App() {
@@ -41,44 +42,10 @@ function App() {
   }, [dispatch]);
 
   // changes to the material ui color palette
-  let theme = createTheme({
-    palette: {
-      primary: {
-        main: '#B8CA30'
-      },
-      secondary: {
-        main: '#17301C',
-      },
-      error: {
-        main: '#ff312e',
-      },
-      warning: {
-        main: '#BF6900',
-      },
-      info: {
-        main: '#3A3335',
-      },
-      success: {
-        main: '#276FBF'
-      },
-      contrastThreshold: 4.5,
-      tonalOffset: 0.5,
-    },
-  });
-
-  theme = createTheme(theme, {
-    // Custom colors created with augmentColor 
-    palette: {
-      pink: theme.palette.augmentColor({
-        color: {
-          main: '#FFEBC8',
-        },
-        name: 'background',
-      }),
-    },
-  });
+  let theme = MaterialTheme();
 
   return (
+    <ThemeProvider theme={theme}>
     <Router>
       <div>
         <Nav />
@@ -216,6 +183,7 @@ function App() {
         <Footer />
       </div>
     </Router>
+    </ThemeProvider>
   );
 }
